requirements.txt
bs4==0.0.1
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
lxml==4.6.3
notebook==6.0.3
numpy==1.21.4
oauthlib==3.2.0
pandas==1.3.5
psycopg2-binary==2.9.3
pylint==2.12.2
PyYAML==5.3.1
regex==2020.11.13
requests
requests-oauthlib==1.3.0
selenium==4.1.5
soupsieve==2.2.1
sqlalchemy==1.4.36
urllib3